Use doubles instead of longs
This change is done because of JMH framework, when we can't exclude
small numbers in measured data. With longs, all such numbers are rounded
to zero and data interpretation fails with exception. Doubles solves
this issue.

+++ src/demo-java/cz/cuni/mff/d3s/spl/demo/SensitivityComparison.java
@@ -175,8 +175,8 @@ public class SensitivityComparison {
 			for (BenchmarkRun run : data.getRuns()) {
 				BenchmarkRunBuilder builder = new BenchmarkRunBuilder();
 				
-				for (long sample : run.getSamples()) {
-					builder.addSamples(Math.round((double)sample * coef));
+				for (double sample : run.getSamples()) {
+					builder.addSamples(Math.round(sample * coef));
 				}
 				
 				snapshotBuilder.addRun(builder.create());
